evaluation/pyproject.toml
[project]
name = "vifood-eval"
version = "0.1.0"
description = "Evaluation scaffold for ViFoodVQA"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datasets>=2.18.0",
    "huggingface-hub>=0.22.0",
    "neo4j>=5.0.0",
    "numpy>=1.24.0",
    "pyarrow>=15.0.0",
    "pillow>=10.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "sentence-transformers>=2.2.0",
]

[project.optional-dependencies]
api = [
    "openai>=1.0.0",
]
hf = [
    "accelerate>=0.27.0",
    "torch>=2.1.0",
    "transformers>=4.51.0",
]
dev = [
    "pytest>=8.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
